Use shared Serial object
Although this is not ideal from a security perspective, for this example
we want to share the Serial object to make sure that the app log does
not get messed up by multiple threads trying to write to the serial
port.

For this purpose, we explicitly instantiated a `shared_pc` Serial
object, instead of silently using `printf` as-is.